Abstract lots of commonality out of the separate concrete types.
Previously, there was a lot of code duplication between the different
variations of the various parsing monads (for state, for laziness, etc).

In this re-organisation, we abstract the 'primitive' operations of each
type into a class PolyParse, defined in Text.ParserCombinators.Poly.Base,
together with many derived combinators which rely only on those primitives.

Then each individual parser type is made an instance of the new class,
each in its own module, defining just the primitives, and re-exporting
the Poly.Base module.

This re-organisation makes it very straightforward to provide ByteString
and Stream versions of the Plain combinator set, where the input tokens
are sequenced in non-list structures, and indeed might not even be
polymorphic (as in ByteString).

The original combinator modules have not been removed in this patch.  I
still want to do some performance measurements to compare the overhead
of the extra class mechanism.