Combinator `apply` is now promoted to a class method.
Apply is basic, because its implementation depends on whether the
underlying parser type is supposed to be lazy or strict.  Many of the
higher-level combinators in T.P.P.Base are now expressed in terms of
`apply`, which removes code duplication in the T.P.P.Lazy module.  In
exchange, `failBad` moves out of the class, because it is always just
(commit . fail).